+++ src/app/search/search.component.ts
@@ -1,9 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { FormControl } from '@angular/forms';
-import { EMPTY } from 'rxjs';
-import { filter, switchMap, debounceTime, catchError } from 'rxjs/operators';
-import { BoardsAPIService } from '../services/boards-api.service';
-import { Board, Task } from '../models/column.model';
+import { filter, debounceTime, distinctUntilChanged } from 'rxjs/operators';
+import { SearchService } from '../services/search.service';
 
 @Component({
   selector: 'app-search-page',
@@ -11,17 +9,9 @@ import { Board, Task } from '../models/column.model';
   styleUrls: ['./search.component.scss'],
 })
 export class SearchComponent implements OnInit {
-  constructor(private reqToBoardsApi: BoardsAPIService) {}
+  constructor(public searchService: SearchService) {}
 
   isBoardSearchVisible = false;
-
-  error: boolean = false;
-
-  user: any = null;
-
-  boards: Board[] = [];
-
-  result: Task[] = [];
 
   findControl = new FormControl();
 
@@ -30,41 +20,14 @@ export class SearchComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    const tokenId = window.localStorage.getItem('userTokenMid');
-    if (tokenId) {
-      this.findControl.valueChanges
-        .pipe(
-          filter((value) => value.length > 2),
-          debounceTime(1000),
-          switchMap(() => {
-            if (this.findControl.value.length > 2) {
-              return this.reqToBoardsApi.getAllBoards(tokenId).pipe(
-                catchError(() => {
-                  this.user = null;
-                  this.error = true;
-                  return EMPTY;
-                }),
-              );
-            }
-            return EMPTY;
-          }),
-        )
-        .subscribe((data) => {
-          this.result = [];
-          this.boards = data;
-          this.boards.forEach((board) => {
-            this.reqToBoardsApi.getBoardByID(board.id, tokenId).subscribe((oneBoardData) => {
-              oneBoardData.columns?.forEach((column: any) => {
-                this.user = column?.tasks.filter((task: any) => task?.title?.includes(this.findControl.value));
-                if (this.user.length !== 0) {
-                  this.result?.push(this.user);
-                }
-              });
-              console.log(this.result);
-            });
-          });
-        });
-      this.error = false;
-    }
+    this.findControl.valueChanges
+      .pipe(
+        filter((value) => value.length > 2),
+        debounceTime(1000),
+        distinctUntilChanged(),
+      )
+      .subscribe((ev) => {
+        if (ev.length > 2) this.searchService.search(ev);
+      });
   }
 }
