+++ src/styles/_variables.scss
@@ -7,6 +7,7 @@ $second-color: #2c2c2cff;
 $fourth-color: #eeeeeeff;
 $error-color: #f95959;
 $bg-loader: #0002;
+$task-color: #009886;
 
 $border-color: #dfdfdfff;
